pyproject.toml
[project]
name = "otp-serivce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "celery>=5.5.3",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]
